+++ src/js/main.js
@@ -1,0 +1,5 @@
+import App from './App';
+
+window.addEventListener('load', () => {
+    new App().render();    
+});