+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# version 1.0.0
+- merge clean-react branch to abandon old setup
+
+# version 0.3.0
+- removed standard js libs
+- added react and react-dom as node modules
+- updated build process
+- cleaned index.html
+
+# version 0.2.0
+- update gulp process to bundle js files
+- added ESLint
+- minified and uglyfied javascript
+- added config file for VSCode
+
 # version 0.1.1
 - added description for css files